requirements.txt
agno>=2.2.10
langchain-openai
browser-use
streamlit
groq
python-dotenv